Resolve the warning witch use unsigned to subtract

+++ src/AccountSystemController.cpp
@@ -94,7 +94,7 @@ nlohmann::json AccountSystemController::run(nlohmann::json &j) {
             // money isn't enough
             success = false;
         } else {
-            accountSystem->modifyMoney(userID, -params[0]);
+            accountSystem->modifyMoney(userID, -(int)params[0]);
             U32vec cardlist;
             if (times > 10) {
                 cardlist.push_back(drawCardSystem->drawCard());
